+++ colamone-ja.html
@@ -71,8 +71,7 @@
     <p  style="text-align:center;font-size:10px;">(c)2014–<script language=JavaScript type=text/javascript> document.write(new Date().getFullYear()); </script> <A Href="https://twitter.com/kurehajime">@kurehajime</A>. All Rights Reserved. / Ver.<a href="https://github.com/kurehajime/colamone_js"  TARGET="_blank" >TIMESTAMP</a>/<span id="time"></span></p>
   </div>
 </div>
-<script type="text/javascript" src="ai.js?TIMESTAMP" defer></script>
-<script type="text/javascript" src="boardgame.js?TIMESTAMP" defer></script>
+<script type="text/javascript" src="main.js?TIMESTAMP" defer></script>
 <script type="text/javascript">
     /** 
      * 国際化・・・しない。
