+++ main.js
@@ -1,0 +1,5 @@
+!function(e){var t={};function n(r){if(t[r])return t[r].exports;var o=t[r]={i:r,l:!1,exports:{}};return e[r].call(o.exports,o,o.exports,n),o.l=!0,o.exports}n.m=e,n.c=t,n.d=function(e,t,r){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:r})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var o in e)n.d(r,o,function(t){return e[t]}.bind(null,o));return r},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=1)}([function(e,t,n){"use strict";
+/* @license Copyright (c) @kurehajime / source code: https://github.com/kurehajime/colamone_js */
+function r(){}n.d(t,"a",function(){return r}),r.copyMap=i,r.thinkAI=function(e,t,n,r,o,l){var u=!1,d=[null,null],s=i(e);l||(l=a);0!==c(s,!1)&&(u=!0);(d=h(s,t,n,r,o,u,l))[1]*t==-999999&&(d=h(s,t,1,r,o,u,l));return d},r.getCanMovePanelX=f,r.isEndX=c,r.isDraw=u;var o=[[0,0,0,0,0,0,0,1,0],[0,1,0,0,0,0,0,1,0],[0,1,0,0,0,0,1,0,1],[1,0,1,0,0,0,1,0,1],[1,0,1,0,0,0,1,1,1],[1,1,1,0,0,0,1,1,1],[1,0,1,1,0,1,1,1,1],[1,1,1,1,0,1,1,1,1],[0,0,0,0,0,0,0,0,0],[1,1,1,1,0,1,1,1,1],[1,1,1,1,0,1,1,0,1],[1,1,1,0,0,0,1,1,1],[1,1,1,0,0,0,1,0,1],[1,0,1,0,0,0,1,0,1],[1,0,1,0,0,0,0,1,0],[0,1,0,0,0,0,0,1,0],[0,1,0,0,0,0,0,0,0]],l=[0,1,2,3,4,5,10,11,12,13,14,15,20,21,22,23,24,25,30,31,32,33,34,35,40,41,42,43,44,45,50,51,52,53,54,55],a={1:[1800,1850,1900,1950,2100,2800],2:[1800,1860,1920,2100,2400,3800],3:[1450,1520,1590,1900,2350,4450],4:[1450,1530,1610,2050,2650,5450],5:[1350,1440,1530,2100,2850,6350],6:[1350,1450,1550,2250,3150,7350],7:[1250,1360,1470,2300,3350,8250],8:[1250,1370,1490,2450,4350,11250]};function i(e){var t=new Int8Array(56);return t[0]=e[0],t[10]=e[10],t[20]=e[20],t[30]=e[30],t[40]=e[40],t[50]=e[50],t[1]=e[1],t[11]=e[11],t[21]=e[21],t[31]=e[31],t[41]=e[41],t[51]=e[51],t[2]=e[2],t[12]=e[12],t[22]=e[22],t[32]=e[32],t[42]=e[42],t[52]=e[52],t[3]=e[3],t[13]=e[13],t[23]=e[23],t[33]=e[33],t[43]=e[43],t[53]=e[53],t[4]=e[4],t[14]=e[14],t[24]=e[24],t[34]=e[34],t[44]=e[44],t[54]=e[54],t[5]=e[5],t[15]=e[15],t[25]=e[25],t[35]=e[35],t[45]=e[45],t[55]=e[55],t}function c(e,t){var n=0,r=0;if(e[0]>0&&(n+=e[0]),e[10]>0&&(n+=e[10]),e[20]>0&&(n+=e[20]),e[30]>0&&(n+=e[30]),e[40]>0&&(n+=e[40]),e[50]>0&&(n+=e[50]),-1*e[5]>0&&(r+=e[5]),-1*e[15]>0&&(r+=e[15]),-1*e[25]>0&&(r+=e[25]),-1*e[35]>0&&(r+=e[35]),-1*e[45]>0&&(r+=e[45]),-1*e[55]>0&&(r+=e[55]),n>=8)return 1;if(r<=-8)return-1;if(d(e))return n>-1*r?1:n<-1*r?-1:0;if(!1===t&&(0!==n||0!==r)){for(var o=0,a=0,i=0;i<=35;i++){var c=0|l[i];e[c]>0?o+=e[c]:e[c]<0&&(a+=e[c])}if(n>-1*(a+r))return 1;if(-1*r>o+n)return-1}return 0}function u(e){if(!d(e))return!1;var t=0,n=0;return e[0]>0&&(t+=e[0]),e[10]>0&&(t+=e[10]),e[20]>0&&(t+=e[20]),e[30]>0&&(t+=e[30]),e[40]>0&&(t+=e[40]),e[50]>0&&(t+=e[50]),-1*e[5]>0&&(n-=e[5]),-1*e[15]>0&&(n-=e[15]),-1*e[25]>0&&(n-=e[25]),-1*e[35]>0&&(n-=e[35]),-1*e[45]>0&&(n-=e[45]),-1*e[55]>0&&(n-=e[55]),t===n}function d(e){for(var t=!1,n=!1,r=0;r<=35;r++){var o=0|l[r];if(0!==e[o])if(!0===s(o,e)&&(e[o]>0?t=!0:e[o]<0&&(n=!0)),t&&n)return!1}return!0}function s(e,t){var n=0|t[e],r=~~(e/10),l=~~(e%10);if(n>0&&0===l||n<0&&5===l||0===n)return!1;for(var a=0;a<9;a++)if(0!==o[n+8][a]){var i=r+~~(a%3)-1,c=l+~~(a/3)-1;if(!(c<0||c>5||i>5||i<0)){var u=t[10*i+c];if(!(u*n>0||u>0&&0===c||u<0&&5===c))return!0}}return!1}function f(e,t){var n=0|t[e],r=~~(e/10),l=~~(e%10),a=[];if(n>0&&0===l||n<0&&5===l||0===n)return a;for(var i=0;i<9;i++)if(0!==o[n+8][i]){var c=r+~~(i%3)-1,u=l+~~(i/3)-1;if(!(u<0||u>5||c>5||c<0)){var d=10*c+u,s=t[d];s*n>0||s>0&&0===u||s<0&&5===u||a.push(d)}}return a}function h(e,t,n,r,o,a,s){var v,m=9999999*t*-1;if(0===n)return[v,m=function(e,t,n,r){var o=0;if(u(e))return 0;var a=c(e,n);if(1===a)return 9999999;if(-1===a)return-9999999;for(var i=0;i<=35;i++){var d=0|l[i],s=0,f=e[d],h=void 0;f>0?(h=5-d%10,s+=r[f][h]):f<0&&(h=d%10,s+=-1*r[-1*f][h]),o+=s}return 0|o}(e,0,a,s)];void 0!==r&&void 0!==o||(r=9999999*t*-1,o=9999999*t);for(var g=function(e,t){for(var n=[],r=0;r<=35;r++){var o=0|l[r];if(!(e[o]*t<=0||0===e[o]))for(var a=f(o,e),c=0;c<a.length;c++){var u=i(e);u[a[c]]=u[o],u[o]=0,n.push([[o,a[c]],u])}}return n}(e,t),p=0;p<g.length;p++){var w=g[p][0],y=g[p][1],S=0,b=c(y,a);if(b===t)return[w,999999*t];if(b!==-1*t){if(S=d(y)?0:h(y,-1*t,n-1,o,r,a,s)[1],void 0===v&&(m=S,v=w),1===t&&S>m?(m=S,v=w):-1===t&&S<m&&(m=S,v=w),(1===t&&r<m||-1===t&&r>m)&&(r=m),1===t&&o<=m||-1===t&&o>=m)break}else void 0===v&&(m=999999*t*-1,v=w)}return[v,m]}},function(e,t,n){"use strict";n.r(t),function(e){var t=n(0);
+/* @license Copyright (c) @kurehajime / source code: https://github.com/kurehajime/colamone_js */!function(e){e.BoardGamejs=function(){},e.BoardGamejs.init=function(){ie(),window.innerHeight<window.innerWidth?document.querySelector(".manual").classList.remove("hide"):document.querySelector(".manual").classList.add("hide");n=document.querySelector("#canv").getContext("2d"),(l=document.createElement("canvas")).width=U,l.height=U,(a=document.createElement("canvas")).width=U,a.height=U,(i=document.createElement("canvas")).width=U,i.height=U,(c=document.createElement("canvas")).width=U,c.height=U,(u=document.createElement("canvas")).width=U,u.height=U,(d=document.createElement("canvas")).width=U,d.height=U,(s=document.createElement("canvas")).width=U,s.height=U,(f=document.createElement("canvas")).width=U,f.height=U,(h=document.createElement("canvas")).width=U,h.height=U,(v=document.createElement("canvas")).width=U,v.height=U,(m=document.createElement("canvas")).width=U,m.height=U,y=U/6,S=1,P=!0,n.canvas.style.width=U/$+"px",n.canvas.style.height=U/$+"px",n.canvas.width=U,n.canvas.height=U,r="ontouchstart"in window;r?(document.querySelector("#canv").addEventListener("touchstart",de),document.querySelector("#canv").addEventListener("touchmove",ue)):(document.querySelector("#canv").addEventListener("mousemove",ce),document.querySelector("#canv").addEventListener("mouseup",de));document.querySelector("#level").addEventListener("change",se),document.querySelector("#prevprev").addEventListener("click",xe),document.querySelector("#prev").addEventListener("click",qe),document.querySelector("#next").addEventListener("click",Pe),document.querySelector("#nextnext").addEventListener("click",Ce),document.querySelector("#replay").addEventListener("click",Ee),document.querySelector("#tweetlog").addEventListener("click",Te),document.querySelector("#newgame").addEventListener("click",Le),document.querySelector("#collapsible").addEventListener("click",function(){document.querySelector(".manual").classList.toggle("hide")}),window.addEventListener("orientationchange",ie),ve(),le.getItem("level_1")||le.setItem("level_1",0);le.getItem("level_2")||le.setItem("level_2",0);le.getItem("level_3")||le.setItem("level_3",0);le.getItem("level_4")||le.setItem("level_4",0);le.getItem("level_5")||le.setItem("level_5",0);void 0!==le.getItem("level_save")&&"undefined"!==le.getItem("level_save")&&null!==le.getItem("level_save")?document.querySelector("#level").value=parseInt(le.getItem("level_save")):(le.setItem("level_save",1),document.querySelector("#level").value=1);var e=Me();e.init?(N=function(e){var n;if(e){for(var r in n=t.a.copyMap(K))n[r]=0;var o=e.split(",");o.length<8&&(o=[1,2,3,4,5,6,7,8]);var l=[0,10,20,30,40,50,11,41],a=[55,45,35,25,15,5,44,14];for(r in a)n[a[r]]=parseInt(o[r]);for(r in l)n[l[r]]=-1*parseInt(o[r])}return n}(e.init),K=t.a.copyMap(N)):N=t.a.copyMap(K);e.log&&(ne=function(e,n){var r=[],o=t.a.copyMap(n),l={q:0,w:1,e:2,a:3,s:4,d:5,z:6,x:7,c:8},a=(e=(e=(e=e.replace(/q/g,"q,").replace(/w/g,"w,").replace(/e/g,"e,")).replace(/a/g,"a,").replace(/s/g,"s,").replace(/d/g,"d,")).replace(/z/g,"z,").replace(/x/g,"x,").replace(/c/g,"c,")).split(",");r.push(o);for(var i=0;i<a.length;i++)if(""!==a[i]){var c=l[a[i].match(/[qweasdzxc]/)[0]],u=parseInt(a[i].match(/\d*/)[0]),d=10*(Math.floor(u/10)+Math.floor(c%3)-1)+(Math.floor(u%10)+Math.floor(c/3)-1);(o=t.a.copyMap(o))[d]=parseInt(o[u]),o[u]=0,r.push(o)}return r}(e.log,N));e.lv&&(document.querySelector("#level").value=parseInt(e.lv));0!==ne.length?(document.querySelector("#log").classList.remove("hide"),document.querySelector("#prevprev").classList.remove("hide"),document.querySelector("#prev").classList.remove("hide"),document.querySelector("#next").classList.remove("hide"),document.querySelector("#nextnext").classList.remove("hide"),document.querySelector("#span_replay").classList.add("hide"),document.querySelector("#span_tweetlog").classList.add("hide"),document.querySelector("#next").focus()):(document.querySelector("#log").classList.add("hide"),document.querySelector("#prevprev").classList.add("hide"),document.querySelector("#prev").classList.add("hide"),document.querySelector("#next").classList.add("hide"),document.querySelector("#nextnext").classList.add("hide"),document.querySelector("#span_replay").classList.add("hide"),document.querySelector("#span_tweetlog").classList.add("hide"));oe.onload=function(){p=!0,ge(!0,!1)},oe.onerror=function(){ge(!0,!1)},0!==oe.width&&(p=!0,ge(!0,!1));setTimeout(function(){0!==oe.width&&(p=!0),ge(!0,!1)},2500),ye(),g=document,w="script",b="twitter-wjs",I=g.getElementsByTagName(w)[0],g.getElementById(b)||((M=g.createElement(w)).id=b,M.async=!0,M.src="https://platform.twitter.com/widgets.js",I.parentNode.insertBefore(M,I)),0===ne.length||(P=!1,C=!0,E=window.setInterval(ae,1e3));var g,w,b,M,I;o=!1,ge(!0,!1)};var n=null,r=!0,o=!1,l=null,a=null,i=null,c=null,u=null,d=null,s=null,f=null,h=null,v=null,m=null,g=!1,p=!1,w=null,y=null,S=null,b=0,M=0,I=null,x="",q=0,P=!0,C=!1,L=null,E=null,T=[],_=0,A="#333333",k="#660033",R="#004466",B="#FFFFFF",O="#111111",X="#444444",F="#888888",j="#7fed7f",Y="#148d14",G="#E60073",H="#0099E6",D="#E60073",z="#0099E6",W="#FFFFFF",$=1;void 0!==window.devicePixelRatio&&1!=window.devicePixelRatio&&($=window.devicePixelRatio);var N,U=500*$,J={1:[1,1,1,1,0,1,1,1,1],2:[1,1,1,1,0,1,1,0,1],3:[1,1,1,0,0,0,1,1,1],4:[1,1,1,0,0,0,1,0,1],5:[1,0,1,0,0,0,1,0,1],6:[1,0,1,0,0,0,0,1,0],7:[0,1,0,0,0,0,0,1,0],8:[0,1,0,0,0,0,0,0,0],"-1":[1,1,1,1,0,1,1,1,1],"-2":[1,0,1,1,0,1,1,1,1],"-3":[1,1,1,0,0,0,1,1,1],"-4":[1,0,1,0,0,0,1,1,1],"-5":[1,0,1,0,0,0,1,0,1],"-6":[0,1,0,0,0,0,1,0,1],"-7":[0,1,0,0,0,0,0,1,0],"-8":[0,0,0,0,0,0,0,1,0]},K=[-1,0,0,0,0,6,0,0,0,0,-2,-8,0,0,7,5,0,0,0,0,-3,0,0,0,0,4,0,0,0,0,-4,0,0,0,0,3,0,0,0,0,-5,-7,0,0,8,2,0,0,0,0,-6,0,0,0,0,1],Q={},V=3,Z=0,ee=0,te=0,ne=[],re=[],oe=null;(oe=new Image).src="bk.gif";var le=null;try{window==parent&&"localStorage"in window&&null!==window.localStorage&&(le=localStorage).setItem("test",0)}catch(e){le=null}function ae(){null!==E&&1==C?Pe():clearInterval(E)}function ie(){var e=document.querySelector("meta[name=viewport]");if(screen.width<500&&screen.height<500?screen.width<screen.height?e.setAttribute("content","width=500,user-scalable=no"):e.setAttribute("content","height=500,user-scalable=no"):screen.width<500?e.setAttribute("content","width=500,user-scalable=no"):screen.height<500&&e.setAttribute("content","height=500,user-scalable=no"),/iPad|iPhone|iPod/.test(navigator.userAgent)&&!window.MSStream){var t=screen.width,n=520;0!==Math.abs(window.orientation)&&(t=screen.height,n=900);var r=Math.round(t/n*1e3)/1e3;r==Math.round(r)&&(r+=1e-4),e.setAttribute("content","initial-scale="+r+", minimum-scale="+r+", maximum-scale="+r+", user-scalable=no")}}function ce(e){me(e),ge(!1,!0)}function ue(e){e.preventDefault(),e.stopPropagation()}function de(e){me(e);var n=10*Math.floor(Z/y)+Math.floor(ee/y);if(null!==I||0!==ne.length)return Le(),!0;if(!0===P)return P=!1,T=[],K=t.a.copyMap(N),re=[],ge(!1,!1),I=null,o=!1,S=1,window.clearInterval(L),ge(!1,!1),!0;if(null===w)K[n]*S>0&&(w=n);else{if(n==w)return w=null,void ge(!1,!1);t.a.getCanMovePanelX(w,K).indexOf(n)>=0&&(ge(!1,!0),he(K,n,S)&&(o=!0,ge(!1,!0),setTimeout(function(){o=!1,ge(!1,!1)},3e3)),K[n]=K[w],K[w]=0,S*=-1,re.push([w,n]),T=[w,n],w=null,x="thinking...",window.setTimeout(function(){ge(!1,!1)},50),ye(),null===I&&window.setTimeout(function(){fe(document.querySelector("#level").value),x="",ye(),ge(!1,!1)},250))}ge(!1,!1)}function se(){var e=document.querySelector("#level").value;le.setItem("level_save",e),le.getItem("level_"+e)>0?document.querySelector("#wins").innerHTML=le.getItem("level_"+e)+" win!":document.querySelector("#wins").innerHTML="",K=t.a.copyMap(N),T=[],Q={},re=[],ge(!1,!1)}function fe(e){var n,r,l=new Date,a=function(e){var n=0;for(var r in e)if(0!==e[r]){var o=t.a.getCanMovePanelX(r,e);n+=o.length}return n}(K)/2,i=0;switch(e=parseInt(e)){case 1:a<=7&&i++;break;case 2:a<=8&&i++;break;case 3:a<=10&&i++,a<=6&&i++;break;case 4:a<=11&&i++,a<=7&&i++;break;case 5:case 6:a>16&&i--,a<=12&&i++,a<=8&&i++}n=t.a.thinkAI(K,S,e+i+1)[0],T=n,n&&(he(K,n[1],S)&&(o=!0,ge(!1,!0),setTimeout(function(){o=!1,ge(!1,!1)},3e3)),K[n[1]]=K[n[0]],K[n[0]]=0,re.push([n[0],n[1]])),S*=-1,r=new Date,q=(r-l)/1e3}function he(e,t,n){if(n>0){if(t%10==0)return!0}else if(n<0&&t%10==5)return!0;return!1}function ve(){for(var e in K)K[e]=0;for(var t=[1,2,3,4,5,6,7,8],n=[0,10,20,30,40,50,11,41],r=[55,45,35,25,15,5,44,14],o=t.length-1;o>=0;o--){var l=Math.floor(Math.random()*(o+1)),a=t[o];t[o]=t[l],t[l]=a}for(e in r)K[r[e]]=t[e];for(e in n)K[n[e]]=-1*t[e]}function me(e){if(null!=e){e.clientX||(e=e.touches?e.touches[0]:e.touches?e.touches[0]:event.touches[0]);var t=e.target.getBoundingClientRect();Z=e.clientX-t.left,ee=e.clientY-t.top,Z*=$,ee*=$}}function ge(e,f){var p=new Int8Array(K);if(n.clearRect(0,0,U,U),!1===f&&(g=!1),!1===f||!1===g){n.drawImage(function(e){if(!1===e)return l;var t=l.getContext("2d");t.clearRect(0,0,U,U);var n=t.createLinearGradient(0,0,U,U);n.addColorStop(0,F),n.addColorStop(.3,X),n.addColorStop(1,O);for(var r=0;r<6;r++)for(var o=0;o<6;o++)t.strokeStyle=A,0===o?t.fillStyle=k:5==o?t.fillStyle=R:(r+o)%2==0?t.fillStyle=B:(t.fillStyle=O,t.fillStyle=n),t.beginPath(),t.fillRect(r*y,o*y,y,y),t.strokeRect(r*y,o*y,y,y);return l}(e),0,0,U,U),n.drawImage(function(e){if(!1===e)return a;var t=a.getContext("2d");return t.clearRect(0,0,U,U),t.globalAlpha=.07,t.fillStyle=W,t.beginPath(),t.arc(1*y,-3*y,7*y,0,2*Math.PI,!1),t.fill(),a}(e),0,0,U,U),null!==w&&(p[w]=0),n.drawImage(function(e,t){var n=u.getContext("2d");n.clearRect(0,0,U,U);var r,o=t[0]/10|0,l=t[0]%10,a=t[1]/10|0,i=t[1]%10,c=y-1*y/5,d=y-1*y/5,s=a*y+y/10,f=i*y+y/10,h=s+d/2,v=f+c/2,m=h,g=v,p=e[t[1]],w="",S=0;if(2!=t.length)return u;w=p>0?H:G;if(o==a||l==i)S=(o+l)%2==0&&0!==l&&5!=l?.5:.3,o<a&&(s=s-d+10,m-=d),o>a&&(s=s+d-10,m+=d),l<i&&(f=f-c+10,g-=c),l>i&&(f=f+c-10,g+=c),(r=n.createLinearGradient(h,v,m,g)).addColorStop(0,w),r.addColorStop(S,w),r.addColorStop(1,"rgba(255, 255, 255, 0)"),n.fillStyle=r,n.fillRect(s,f,c,d);else{var b,M,I,x,q,P,C,L;S=(o+l)%2==0&&0!==l&&5!=l?.3:.5,o<a&&l<i&&(b=s,M=f-c,I=s+d,x=f,q=s,P=f+c,C=s-d,L=f,m-=d,g-=c,h+=2*d,v+=2*c),o>a&&l>i&&(b=s+d,M=f,I=s+d+d,x=f+c,q=s+d,P=f+c+c,C=s,L=f+c,m+=d,g+=c,h-=2*d,v-=2*c),o<a&&l>i&&(b=s,M=f,I=s+d,x=f+c,q=s,P=f+c+c,C=s-d,L=f+c,m-=d,g+=c,h+=2*d,v-=2*c),o>a&&l<i&&(b=s+d,M=f-c,I=s+d+d,x=f,q=s+d,P=f+c,C=s,L=f,m+=d,g-=c,h-=2*d,v+=2*c),n.beginPath(),n.moveTo(b,M),n.lineTo(I,x),n.lineTo(q,P),n.lineTo(C,L),n.lineTo(b,M),n.closePath(),(r=n.createLinearGradient(h,v,m,g)).addColorStop(0,w),r.addColorStop(S,w),r.addColorStop(.97,"rgba(255, 255, 255, 0)"),n.fillStyle=r,n.fill()}return u}(p,T),0,0,U,U),n.drawImage(function(e){var t=c.getContext("2d");t.clearRect(0,0,U,U);for(var n=0;n<6;n++)for(var r=0;r<6;r++)if(0!==e[10*n+r]){var o=!1;e[10*n+r]>0&&0===r?o=!0:e[10*n+r]<0&&5==r&&(o=!0),t=pe(t,n*y,r*y,e[10*n+r],o)}return c}(p),0,0,U,U);var q=m.getContext("2d");q.clearRect(0,0,U,U),q.drawImage(n.canvas,0,0,U,U),g=!0}else n.drawImage(m,0,0,U,U);n.drawImage(function(){var e=d.getContext("2d");e.clearRect(0,0,U,U);var t=Z-y/2,n=ee-y/2;null!==w&&pe(e,t,n,K[w],!1);return d}(),0,0,U,U),0===Z&&0===ee||!1!==P||n.drawImage(function(){var e=Z-Z%y,n=ee-ee%y,o=i.getContext("2d"),l=o.createRadialGradient(e,n,0,e,n,y);if(l.addColorStop(.3,j),l.addColorStop(1,Y),o.clearRect(0,0,U,U),o.globalAlpha=.35,o.fillStyle=l,o.lineWidth=1,o.beginPath(),o.fillRect(e,n,y,y),!0===r&&null===w)return i;var a=e/y*10+n/y;if(K[a]*S>0)for(var c=t.a.getCanMovePanelX(a,K),u=0;u<=c.length-1;u++)e=Math.floor(c[u]/10),n=Math.floor(c[u]%10),o.globalAlpha=.6,o.strokeStyle=j,o.lineWidth=y/20,o.beginPath(),o.arc(e*y+y/2,n*y+y/2,y/2-y/10,0,2*Math.PI,!1),o.stroke();return i}(),0,0,U,U),(o||null!==I)&&!1===P&&0==C&&n.drawImage(function(){var e=v.getContext("2d"),t="",n=Math.round(1.5*y),r=z,o=D;return e.clearRect(0,0,U,U),e.globalAlpha=.4,e.textBaseline="middle",e.textAlign="center",e.shadowBlur=10,e.shadowColor="rgba(0, 0, 0, 0.5)",e.font="bold "+n+"px Play,sans-serif",e.globalAlpha=.3,e.shadowBlur=2,e.shadowColor="rgba(255, 255, 255, 1)",e.fillStyle=o,t=M,e.beginPath(),e.fillText(t,1*y,3.8*y),t="8",e.beginPath(),e.fillText(t,2*y,5.3*y),e.lineWidth=.2*y,e.strokeStyle=o,e.beginPath(),e.moveTo(.4*y,5.55*y),e.lineTo(2.6*y,3.55*y),e.closePath(),e.stroke(),t=b,e.fillStyle=r,e.beginPath(),e.fillText(t,4*y,.7*y),t="8",e.beginPath(),e.fillText(t,5*y,2.3*y),e.lineWidth=.2*y,e.strokeStyle=r,e.beginPath(),e.moveTo(3.4*y,2.55*y),e.lineTo(5.6*y,.55*y),e.closePath(),e.stroke(),v}(),0,0,U,U),n.drawImage(function(){var e=s.getContext("2d"),t=1.3*y,n=2.5*y;if(e.clearRect(0,0,U,U),""===x)return s;e.shadowBlur=10,e.shadowColor="rgba(100, 100, 100, 0.5)",e.shadowOffsetX=5,e.shadowOffsetY=5,e.globalAlpha=.9,e.fillStyle=W,we(e,t,n,3.4*y,1*y,y/20);var r=Math.round(.36*y);return e.shadowBlur=0,e.shadowOffsetX=0,e.shadowOffsetY=0,e.shadowColor="rgba(0, 0, 0, 0)",e.font="bold "+r+"px Play,sans-serif",e.globalAlpha=1,e.fillStyle=A,e.textBaseline="middle",e.textAlign="center",e.beginPath(),e.fillText(x,3*y,3*y),s}(),0,0,U,U),!0===P&&n.drawImage(function(){var e=h.getContext("2d");e.clearRect(0,0,U,U),e.globalAlpha=.5,e.fillStyle="#000000",e.fillRect(0,0,U,U);var t=2*y,n=3.5*y;e.shadowBlur=20,e.shadowColor="rgba(0, 0, 0, 0.3)",e.shadowOffsetX=5,e.shadowOffsetY=5,e.globalAlpha=.8,e.fillStyle=W,e.beginPath(),we(e,t,n,2*y,1*y,y/20),e.shadowColor="rgba(0, 0, 0, 0)",e.shadowBlur=0,e.shadowOffsetX=0,e.shadowOffsetY=0;var r=Math.round(.5*y),o="Play";return o+=_%10==0?" ":"",e.shadowBlur=0,e.shadowOffsetX=0,e.shadowOffsetY=0,e.shadowColor="rgba(0, 0, 0, 0)",e.font="bold "+r+"px Play,sans-serif",e.globalAlpha=1,e.fillStyle=A,e.textBaseline="middle",e.textAlign="center",e.beginPath(),e.fillText(o,3*y,4*y),o="colamone",r=Math.round(1*y),e.font="bold "+r+"px Play,sans-serif",e.fillStyle=W,e.shadowBlur=0,e.beginPath(),e.fillText(o,3*y,2*y),h}(),0,0,U,U)}function pe(e,t,r,o,l){var a;if(0===o)return e;a=o>0?H:G;var i,c=n.createLinearGradient(t,r,t+y,r+y);c.addColorStop(0,"rgb(255, 255, 255)"),c.addColorStop(.4,a),c.addColorStop(1,a),e.shadowBlur=10,e.shadowColor="rgba(0, 0, 0, 1)",e.shadowOffsetX=2,e.shadowOffsetY=2,e.fillStyle=c,e.beginPath(),we(e,t+y/10,r+y/10,y-1*y/5,y-1*y/5,y/20),e.shadowColor="rgba(0, 0, 0, 0)",e.shadowBlur=0,e.shadowOffsetX=0,e.shadowOffsetY=0,p&&(e.globalAlpha=.35,e.save(),e.clip(),e.drawImage((i=f.getContext("2d"),p&&i.drawImage(oe,0,0,U/$,U/$,0,0,U,U),f),t+y/10,r+y/10,y-1*y/5,y-1*y/5),e.restore(),e.globalAlpha=1),e.fillStyle=W;var u=Math.round(.18*y);e.textBaseline="middle",e.textAlign="center",e.font=u+"pt 'Play',Arial",e.beginPath(),e.fillText(Math.abs(o),t+y/2,r+y/2);for(var d=0;d<=J[o].length-1;d++)if(0!==J[o][d]){var s=t+y/4.16+Math.floor(y-1*y/5)/3*Math.floor(d%3),h=r+y/4.16+Math.floor(y-1*y/5)/3*Math.floor(d/3);e.fillStyle=W,e.beginPath(),e.arc(s,h,.06*y,0,2*Math.PI,!1),e.fill()}return l&&(e.shadowBlur=10,e.shadowColor="rgba(0, 0, 0, 1)",e.globalAlpha=1,e.fillStyle=W,u=Math.round(.5*y),e.textBaseline="middle",e.textAlign="center",e.font="bolder "+u+"pt Play,Arial ",e.beginPath(),e.fillText(Math.abs(o),t+y/2,r+y/2),e.globalAlpha=1,e.shadowColor="rgba(0, 0, 0, 0)",e.shadowBlur=0),e}function we(e,t,n,r,o,l){e.beginPath(),e.moveTo(t+l,n),e.lineTo(t+r-l,n),e.arc(t+r-l,n+l,l,1.5*Math.PI,0,!1),e.lineTo(t+r,n+o-l),e.arc(t+r-l,n+o-l,l,0,.5*Math.PI,!1),e.lineTo(t+l,n+o),e.arc(t+l,n+o-l,l,.5*Math.PI,Math.PI,!1),e.lineTo(t,n+l),e.arc(t+l,n+l,l,Math.PI,1.5*Math.PI,!1),e.closePath(),e.fill()}function ye(){be();document.querySelector("#blue").innerHTML="Blue: "+b+"/8",document.querySelector("#red").innerHTML=" Red: "+M+"/8",document.querySelector("#time").innerHTML="("+q.toFixed(3)+"sec)",0===ne.length&&(1==I?(x="You win!",le.setItem("level_"+document.querySelector("#level").value,parseInt(le.getItem("level_"+document.querySelector("#level").value))+1),Se()):-1==I?(x="You lose...",le.setItem("level_"+document.querySelector("#level").value,0),Se()):0===I&&(x=Q[JSON.stringify(K)]>=V?"3fold repetition":"-- Draw --",Se())),le.getItem("level_"+document.querySelector("#level").value)>0?document.querySelector("#wins").innerHTML=le.getItem("level_"+document.querySelector("#level").value)+" win!":document.querySelector("#wins").innerHTML=""}function Se(){0===ne.length&&(document.querySelector("#span_replay").classList.remove("hide"),document.querySelector("#span_tweetlog").classList.remove("hide"))}function be(){var e=0,n=0,r=[0,10,20,30,40,50],o=[5,15,25,35,45,55];for(var l in r)1*K[r[l]]>0&&(e+=K[r[l]]);for(l in o)-1*K[o[l]]>0&&(n+=K[o[l]]);e>=8?I=1:n<=-8&&(I=-1),!function(e){var n=!1,r=!1;for(var o in e)if(0!==e[o]){var l=t.a.getCanMovePanelX(o,e);if(0!==l.length&&(e[o]>0?n=!0:e[o]<0&&(r=!0)),n&&r)return!1}return!0}(K)?!0===function(e){var t=JSON.stringify(e);if(void 0===Q[t])return Q[t]=1,!1;Q[t]+=1;if(Q[t]>=V)return!0;return!1}(K)&&(I=0):Math.abs(e)>Math.abs(n)?I=1:Math.abs(e)<Math.abs(n)?I=-1:Math.abs(e)==Math.abs(n)&&(I=0),b=Math.abs(e),M=Math.abs(n)}function Me(){var e={};if(1<document.location.search.length)for(var t=document.location.search.substring(1).split("&"),n=0;n<t.length;n++){var r=t[n].split("="),o=decodeURIComponent(r[0]),l=decodeURIComponent(r[1]);e[o]=decodeURIComponent(l)}return e}function Ie(e){var t="";for(var n in e){var r=e[n][0],o=e[n][1],l=Math.floor(o/10)-Math.floor(r/10),a=Math.floor(o%10)-Math.floor(r%10),i="";-1===l&&-1===a&&(i="q"),0===l&&-1===a&&(i="w"),1===l&&-1===a&&(i="e"),-1===l&&0===a&&(i="a"),0===l&&0===a&&(i="s"),1===l&&0===a&&(i="d"),-1===l&&1===a&&(i="z"),0===l&&1===a&&(i="x"),1===l&&1===a&&(i="c"),t+=r+i}return t}function xe(){te=0,C=!1,K=t.a.copyMap(ne[te]),I=null,o=!1,ye(),ge(!1,!1)}function qe(){te<=0||(C=!1,te-=1,K=t.a.copyMap(ne[te]),I=null,o=!1,ye(),ge(!1,!1))}function Pe(){te+1>ne.length-1||(te+=1,K=t.a.copyMap(ne[te]),ye(),ge(!1,!1))}function Ce(){te=ne.length-1,C=!1,K=t.a.copyMap(ne[te]),ye(),ge(!1,!1)}function Le(){var e=document.location.href.split("?")[0];if(!0!==P){var n=Me();n.lang&&(e+="?lang="+n.lang),navigator.onLine?location.href=e:(K=t.a.copyMap(N),ve(),re=[],x="",I=null,S=1,ge(!1,!1))}else de(null)}function Ee(){var e=document.location.href.split("?")[0],t="?init="+N[55]+","+N[45]+","+N[35]+","+N[25]+","+N[15]+","+N[5]+","+N[44]+","+N[14],n="&log="+Ie(re);n+="&lv="+document.querySelector("#level").value,location.href=e+t+n}function Te(){var e=document.location.href.split("?")[0],t="?init="+N[55]+","+N[45]+","+N[35]+","+N[25]+","+N[15]+","+N[5]+","+N[44]+","+N[14],n="%26log="+Ie(re);n+="%26lv="+document.querySelector("#level").value,window.open("https://twitter.com/intent/tweet?text="+e+t+n+"%20%23colamone")}null===le&&((le={}).getItem=function(){},le.setItem=function(){},navigator.cookieEnabled&&(le.hasItem=function(e){return new RegExp("(?:^|;\\s*)"+escape(e).replace(/[\-\.\+\*]/g,"\\$&")+"\\s*\\=").test(document.cookie)},le.getItem=function(e){return e&&new RegExp("(?:^|;\\s*)"+escape(e).replace(/[\-\.\+\*]/g,"\\$&")+"\\s*\\=").test(document.cookie)?unescape(document.cookie.replace(new RegExp("(?:^|.*;\\s*)"+escape(e).replace(/[\-\.\+\*]/g,"\\$&")+"\\s*\\=\\s*((?:[^;](?!;))*[^;]?).*"),"$1")):null},le.setItem=function(e,t){e&&!/^(?:expires|max\-age|path|domain|secure)$/i.test(e)&&(document.cookie=escape(e)+"="+escape(t))}))}((0).self||e),document.addEventListener("DOMContentLoaded",function(){BoardGamejs.init()})}.call(this,n(2))},function(e,t){function n(e){return(n="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}var r;r=function(){return this}();try{r=r||new Function("return this")()}catch(e){"object"===("undefined"==typeof window?"undefined":n(window))&&(r=window)}e.exports=r}]);
+//# sourceMappingURL=main.js.map
