+++ package.json
@@ -1,0 +1,31 @@
+{
+  "name": "colamone_js",
+  "version": "1.0.0",
+  "description": "Colamone - A two-player strategy board game ===========",
+  "main": "main.js",
+  "directories": {
+    "lib": "lib"
+  },
+  "scripts": {
+    "build": "webpack",
+    "watch": "webpack --watch"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/kurehajime/colamone_js.git"
+  },
+  "keywords": [],
+  "author": "kurehajime",
+  "license": "",
+  "bugs": {
+    "url": "https://github.com/kurehajime/colamone_js/issues"
+  },
+  "homepage": "https://xiidec.appspot.com/colamone/colamone.html",
+  "devDependencies": {
+    "@babel/core": "^7.4.5",
+    "@babel/preset-env": "^7.4.5",
+    "babel-loader": "^8.0.6",
+    "webpack": "^4.33.0",
+    "webpack-cli": "^3.3.3"
+  }
+}
